secrets-scanner/orchestrator/pyproject.toml
[project]
name = "secrets-orchestrator"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "requests>=2.31.0",
    "gitpython>=3.1.40",
    "gunicorn>=21.0.0",
    "google-genai>=0.2.0",
]
